pyproject.toml
[project]
name = "nert"
version = "0.1.0"
description = "NER Knowledge Application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gliner>=0.2.25",
    "streamlit>=1.55.0",
    "sqlalchemy>=2.0.48",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "wordcloud>=1.9.6",
    "st-annotated-text>=4.0.2",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "requests>=2.32.5",
    "pydantic>=2.12.5",
    "ollama>=0.6.1",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/db", "backend/services"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
